Moved `soft_escape` back into `AU2.plugins.util.game`
(Also moved `HTML_REPORT_PREFIX`)

This is because having it in `render_utils` caused a dependency cycle.
